Build-F: Fixing warning in windows build

+++ src/raylib/Text.cpp
@@ -7,17 +7,17 @@
 
 #include "Text.hpp"
 
-void raylib::drawText(const std::string &text, int posX, int posY, float fontSize, ::Color tint)
+void raylib::drawText(const std::string &text, int posX, int posY, int fontSize, ::Color tint)
 {
     ::DrawText(text.c_str(), posX, posY, fontSize, tint);
 }
 
-void raylib::drawText(::Font &font, const std::string &text, ::Vector2 position, int fontSize, ::Color tint)
+void raylib::drawText(::Font &font, const std::string &text, ::Vector2 position, float fontSize, ::Color tint)
 {
     ::DrawTextEx(font, text.c_str(), position, fontSize, 2, tint);
 }
 
-int raylib::measureText(const std::string &text, float fontSize)
+int raylib::measureText(const std::string &text, int fontSize)
 {
     return (::MeasureText(text.c_str(), fontSize));
 }
